`imp_push_get_ref` is implemented
Consider the following pattern of calling

`vec.imp_push(value);`

followed by

`&vec[vec.len() - 1]`

to get the reference to the pushed element. This is a frequently used pattern with `ImpVec` since it reflects its core functionality.

Therefore, the two calls are wrapped in the `imp_push_get_ref` function which also allows the caller not to deal with indices.

In addition, three more examples are added to the repository to demonstrate the motivation of the crate.